lib ok sans bonus

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: anporced <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 14:57:56 by anporced          #+#    #+#             */
-/*   Updated: 2023/10/04 20:45:06 by anporced         ###   ########.fr       */
+/*   Updated: 2023/10/09 23:21:31 by anporced         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,18 +14,18 @@
 
 char	*ft_strnstr(const char *s, const char *to_find, size_t len)
 {
-	int		i;
-	int		j;
-	int		k;
+	size_t	i;
+	size_t	j;
+	size_t	k;
 
 	i = 0;
 	if (to_find[0] == '\0')
 		return ((char *)s);
-	while (s[i] && i < (int)len)
+	while (s[i] && i < len)
 	{
 		j = 0;
 		k = i;
-		while (s[k] == to_find[j] && k < (int)len)
+		while (s[k] == to_find[j] && k < len)
 		{
 			k++;
 			j++;
